--- src/main/java/common/EncryptedMessage.java
@@ -1,6 +1,0 @@
-package common;
-
-public class EncryptedMessage {
-    public byte[] iv;
-    public byte[] ciphertext;
-}
