--- src/main/java/common/KeyExchange.java
@@ -1,7 +1,0 @@
-package common;
-
-public class KeyExchange {
-    public byte[] encryptedKey;
-}
-
-
