--- src/main/java/common/LoginRequest.java
@@ -1,6 +1,0 @@
-package common;
-
-public class LoginRequest {
-    public String username;
-    public String password;
-}
