--- src/main/java/common/LoginResponse.java
@@ -1,6 +1,0 @@
-package common;
-
-public class LoginResponse {
-    public boolean success;
-    public String message;
-}
