--- src/main/java/common/Message.java
@@ -1,5 +1,0 @@
-package common;
-
-public class Message {
-    public String text;
-}
