--- src/main/java/common/RegisterRequest.java
@@ -1,6 +1,0 @@
-package common;
-
-public class RegisterRequest {
-    public String username;
-    public String password;
-}
