--- src/main/java/common/RegisterResponse.java
@@ -1,6 +1,0 @@
-package common;
-
-public class RegisterResponse {
-    public boolean success;
-    public String message;
-}
