+++ src/main/java/networked/messages/EncryptedMessage.java
@@ -1,0 +1,6 @@
+package networked.messages;
+
+public class EncryptedMessage extends PreSessionMessage {
+    public byte[] iv;
+    public byte[] ciphertext;
+}
