+++ src/main/java/networked/messages/KeyExchangeMessage.java
@@ -1,0 +1,7 @@
+package networked.messages;
+
+public class KeyExchangeMessage extends PreSessionMessage {
+    public byte[] encryptedKey;
+}
+
+
