+++ src/main/java/networked/messages/LoginRequest.java
@@ -1,0 +1,6 @@
+package networked.messages;
+
+public class LoginRequest extends NetworkedMessage {
+    public String username;
+    public String password;
+}
