+++ src/main/java/networked/messages/LoginResponse.java
@@ -1,0 +1,6 @@
+package networked.messages;
+
+public class LoginResponse extends NetworkedMessage {
+    public boolean success;
+    public String message;
+}
