+++ src/main/java/networked/messages/RegisterRequest.java
@@ -1,0 +1,6 @@
+package networked.messages;
+
+public class RegisterRequest extends NetworkedMessage {
+    public String username;
+    public String password;
+}
