+++ src/main/java/networked/messages/RegisterResponse.java
@@ -1,0 +1,6 @@
+package networked.messages;
+
+public class RegisterResponse extends NetworkedMessage {
+    public boolean success;
+    public String message;
+}
